+++ packages/frontend/src/components/endpoints/EndpointState.tsx
@@ -34,10 +34,10 @@ export const EndpointState = (props: EndpointStateProps) => {
   );
   const metadata = useMemo(
     () => ({
-      "Endpoint ID": props.endpoint.id.local,
-      "Endpoint Type": `${props.endpoint.type.name} (${props.endpoint.type.id})`,
-      "Endpoint Number": props.endpoint.endpoint,
-      "# of Child Endpoints": props.endpoint.parts.length,
+      "端点 ID": props.endpoint.id.local,
+      端点类型: `${props.endpoint.type.name} (${props.endpoint.type.id})`,
+      端点编号: props.endpoint.endpoint,
+      子端点数量: props.endpoint.parts.length,
     }),
     [props.endpoint],
   );
@@ -46,7 +46,7 @@ export const EndpointState = (props: EndpointStateProps) => {
     <>
       <Paper sx={{ p: 2, mb: 2 }} variant="outlined">
         <Stack spacing={2}>
-          <Typography component="span">About this endpoint</Typography>
+          <Typography component="span">关于此端点</Typography>
           <ObjectTable value={metadata} hideHead></ObjectTable>
         </Stack>
       </Paper>
@@ -58,7 +58,7 @@ export const EndpointState = (props: EndpointStateProps) => {
             aria-controls="panel1-content"
           >
             <Typography component="span">
-              Behavior: <strong>{behavior}</strong>
+              行为：<strong>{behavior}</strong>
             </Typography>
           </AccordionSummary>
           <AccordionDetails>
@@ -84,8 +84,8 @@ const ObjectTable = <T extends object>(props: {
         {!props.hideHead && (
           <TableHead>
             <TableRow>
-              <TableCell>Property</TableCell>
-              <TableCell>Value</TableCell>
+              <TableCell>属性</TableCell>
+              <TableCell>值</TableCell>
             </TableRow>
           </TableHead>
         )}